+++ src/libtrx/game/lara/cheat.c
@@ -1,10 +1,12 @@
 #include "game/lara/cheat.h"
 
 #include "game/console.h"
+#include "game/const.h"
 #include "game/game.h"
 #include "game/game_string.h"
 #include "game/inventory.h"
 #include "game/lara.h"
+#include "game/objects.h"
 #include "game/sound.h"
 
 static void M_GiveAllKeysImpl(void);
@@ -121,4 +123,54 @@ bool Lara_Cheat_KillEnemy(const int16_t item_num)
     Sound_Effect(SFX_EXPLOSION_CHEAT, &item->pos, SPM_NORMAL);
     Creature_Die(item_num, true);
     return true;
+}
+
+bool Lara_Cheat_OpenNearestDoor(void)
+{
+    const ITEM *const lara_item = Lara_GetItem();
+    if (lara_item == nullptr) {
+        return false;
+    }
+
+    int32_t opened = 0;
+    int32_t closed = 0;
+
+    const int32_t shift = 8; // constant shift to avoid overflow errors
+    const int32_t max_dist = SQUARE((WALL_L * 2) >> shift);
+    for (int32_t item_num = 0; item_num < Item_GetLevelCount(); item_num++) {
+        ITEM *const item = Item_Get(item_num);
+        if (!Object_IsType(item->object_id, g_DoorObjects)
+            && !Object_IsType(item->object_id, g_TrapdoorObjects)) {
+            continue;
+        }
+
+        const int32_t dx = (item->pos.x - lara_item->pos.x) >> shift;
+        const int32_t dy = (item->pos.y - lara_item->pos.y) >> shift;
+        const int32_t dz = (item->pos.z - lara_item->pos.z) >> shift;
+        const int32_t dist = SQUARE(dx) + SQUARE(dy) + SQUARE(dz);
+        if (dist > max_dist) {
+            continue;
+        }
+
+        if (!item->active) {
+            Item_AddActive(item_num);
+            item->flags |= IF_CODE_BITS;
+            opened++;
+        } else if ((item->flags & IF_CODE_BITS) != 0) {
+            item->flags &= ~IF_CODE_BITS;
+            closed++;
+        } else {
+            item->flags |= IF_CODE_BITS;
+            opened++;
+        }
+        item->timer = 0;
+        item->touch_bits = 0;
+    }
+
+    if (opened > 0 || closed > 0) {
+        Console_Log(opened > 0 ? GS(OSD_DOOR_OPEN) : GS(OSD_DOOR_CLOSE));
+        return true;
+    }
+    Console_Log(GS(OSD_DOOR_OPEN_FAIL));
+    return false;
 }
