+++ src/libtrx/include/libtrx/game/game_string.def
@@ -69,6 +69,9 @@ GS_DEFINE(OSD_GIVE_ITEM_ALL_KEYS, "Surprise! Every key item Lara needs is now in
 GS_DEFINE(OSD_GIVE_ITEM_ALL_GUNS, "Lock'n'load - Lara's armed to the teeth!")
 GS_DEFINE(OSD_GIVE_ITEM_CHEAT, "Lara's backpack just got way heavier!")
 GS_DEFINE(OSD_COMPLETE_LEVEL, "Level complete!")
+GS_DEFINE(OSD_DOOR_OPEN, "Open Sesame!")
+GS_DEFINE(OSD_DOOR_CLOSE, "Close Sesame!")
+GS_DEFINE(OSD_DOOR_OPEN_FAIL, "No doors in Lara's proximity")
 GS_DEFINE(CONTROLS_DEFAULT_KEYS, "Default Keys")
 GS_DEFINE(CONTROLS_CUSTOM_1, "User Keys 1")
 GS_DEFINE(CONTROLS_CUSTOM_2, "User Keys 2")
