+++ src/libtrx/include/libtrx/game/objects/vars.h
@@ -9,6 +9,8 @@ extern const GAME_OBJECT_ID g_AllyObjects[];
 extern const GAME_OBJECT_ID g_PickupObjects[];
 extern const GAME_OBJECT_ID g_SwitchObjects[];
 extern const GAME_OBJECT_ID g_GunObjects[];
+extern const GAME_OBJECT_ID g_DoorObjects[];
+extern const GAME_OBJECT_ID g_TrapdoorObjects[];
 extern const GAME_OBJECT_ID g_AnimObjects[];
 extern const GAME_OBJECT_ID g_NullObjects[];
 extern const GAME_OBJECT_ID g_InvObjects[];
