+++ src/tr1/game/game_string.def
@@ -24,9 +24,6 @@ GS_DEFINE(OSD_FLY_MODE_OFF, "Fly mode disabled")
 GS_DEFINE(OSD_TEXTURE_FILTER_SET, "Texture filter set to %s")
 GS_DEFINE(OSD_TEXTURE_FILTER_NN, "nearest-neighbor")
 GS_DEFINE(OSD_TEXTURE_FILTER_BILINEAR, "bilinear")
-GS_DEFINE(OSD_DOOR_OPEN, "Open Sesame!")
-GS_DEFINE(OSD_DOOR_CLOSE, "Close Sesame!")
-GS_DEFINE(OSD_DOOR_OPEN_FAIL, "No doors in Lara's proximity")
 GS_DEFINE(MISC_EMPTY_SLOT_FMT, "- EMPTY SLOT %d -")
 
 GS_DEFINE(CONSOLE_HELP_DEBUG, "Toggles visual debug information.")
