+++ src/tr1/game/objects/vars.h
@@ -1,6 +1,3 @@
 #pragma once
 
 #include <libtrx/game/objects/vars.h>
-
-extern const GAME_OBJECT_ID g_DoorObjects[];
-extern const GAME_OBJECT_ID g_TrapdoorObjects[];
