+++ src/tr2/game/game_string.def
@@ -4,9 +4,6 @@ GS_DEFINE(OSD_FLIPMAP_ON, "Flipmap set to ON")
 GS_DEFINE(OSD_FLIPMAP_OFF, "Flipmap set to OFF")
 GS_DEFINE(OSD_FLIPMAP_FAIL_ALREADY_ON, "Flipmap is already ON")
 GS_DEFINE(OSD_FLIPMAP_FAIL_ALREADY_OFF, "Flipmap is already OFF")
-GS_DEFINE(OSD_DOOR_OPEN, "Open Sesame!")
-GS_DEFINE(OSD_DOOR_CLOSE, "Close Sesame!")
-GS_DEFINE(OSD_DOOR_OPEN_FAIL, "No doors in Lara's proximity")
 GS_DEFINE(OSD_LOAD_GAME, "Loaded game from save slot %d")
 GS_DEFINE(OSD_LOAD_GAME_FAIL_UNAVAILABLE_SLOT, "Save slot %d is not available")
 GS_DEFINE(OSD_SAVE_GAME, "Saved game to save slot %d")
