+++ src/tr2/game/lara/cheat.c
@@ -159,55 +159,6 @@ bool Lara_Cheat_ExitFlyMode(void)
     return true;
 }
 
-bool Lara_Cheat_OpenNearestDoor(void)
-{
-    if (g_LaraItem == nullptr) {
-        return false;
-    }
-
-    int32_t opened = 0;
-    int32_t closed = 0;
-
-    const int32_t shift = 8; // constant shift to avoid overflow errors
-    const int32_t max_dist = SQUARE((WALL_L * 2) >> shift);
-    for (int32_t item_num = 0; item_num < Item_GetLevelCount(); item_num++) {
-        ITEM *const item = Item_Get(item_num);
-        if (!Object_IsType(item->object_id, g_DoorObjects)
-            && !Object_IsType(item->object_id, g_TrapdoorObjects)) {
-            continue;
-        }
-
-        const int32_t dx = (item->pos.x - g_LaraItem->pos.x) >> shift;
-        const int32_t dy = (item->pos.y - g_LaraItem->pos.y) >> shift;
-        const int32_t dz = (item->pos.z - g_LaraItem->pos.z) >> shift;
-        const int32_t dist = SQUARE(dx) + SQUARE(dy) + SQUARE(dz);
-        if (dist > max_dist) {
-            continue;
-        }
-
-        if (!item->active) {
-            Item_AddActive(item_num);
-            item->flags |= IF_CODE_BITS;
-            opened++;
-        } else if (item->flags & IF_CODE_BITS) {
-            item->flags &= ~IF_CODE_BITS;
-            closed++;
-        } else {
-            item->flags |= IF_CODE_BITS;
-            opened++;
-        }
-        item->timer = 0;
-        item->touch_bits = 0;
-    }
-
-    if (opened > 0 || closed > 0) {
-        Console_Log(opened > 0 ? GS(OSD_DOOR_OPEN) : GS(OSD_DOOR_CLOSE));
-        return true;
-    }
-    Console_Log(GS(OSD_DOOR_OPEN_FAIL));
-    return false;
-}
-
 bool Lara_Cheat_Teleport(int32_t x, int32_t y, int32_t z, int16_t room_num)
 {
     if (room_num == NO_ROOM) {
