+++ src/tr2/game/lara/cheat.h
@@ -4,4 +4,3 @@
 
 bool Lara_Cheat_EnterFlyMode(void);
 bool Lara_Cheat_ExitFlyMode(void);
-bool Lara_Cheat_OpenNearestDoor(void);
