+++ src/tr2/game/objects/vars.h
@@ -2,7 +2,5 @@
 
 #include <libtrx/game/objects/vars.h>
 
-extern const GAME_OBJECT_ID g_DoorObjects[];
-extern const GAME_OBJECT_ID g_TrapdoorObjects[];
 extern const GAME_OBJECT_ID g_SecretObjects[];
 extern const GAME_OBJECT_ID g_SmashableObjects[];
